Refactor PokemonService to use SvcFilterService for filtering
Refactor the `PokemonService` to use the `SvcFilterService` for filtering the list of Pokemon. This involves updating the references to the `pokemonService` to `svcFilterService` in the `filters.component.html` and `filters.component.ts` files. By using the `SvcFilterService`, we ensure that the filtered options are correctly updated and displayed in the autocomplete and checkbox components.